+++ src/app/Fake.IO.FileSystem/GlobbingFileSystem.fs
@@ -1,5 +1,5 @@
 /// This module contains a file pattern globbing implementation.
-namespace Fake.Core
+namespace Fake.IO
 open System.Collections.Generic
 
 type IGlobbingPattern =
@@ -8,8 +8,8 @@ type IGlobbingPattern =
     abstract Includes : string list
     abstract Excludes : string list
 
-namespace Fake.Core.Globbing
-open Fake.Core
+namespace Fake.IO.Globbing
+open Fake.IO
 open System.Collections.Generic
 
 type LazyGlobbingPattern =
@@ -61,9 +61,9 @@ type ResolvedGlobbingPattern =
         member this.GetEnumerator() = (this.Results :> IEnumerable<string>).GetEnumerator()
         member this.GetEnumerator() = (this :> IEnumerable<string>).GetEnumerator() :> System.Collections.IEnumerator
 
-namespace Fake.Core
+namespace Fake.IO
 open System.IO
-open Fake.Core.Globbing
+open Fake.IO.Globbing
 
 [<AutoOpen>] // A bit of a hack but we need those extensions for backwards compat.
 module GlobbingPatternExtensions =
@@ -142,9 +142,9 @@ module GlobbingPattern =
                             |> not)
             |> Seq.toList
 
-namespace Fake.Core.Globbing
+namespace Fake.IO.Globbing
 
-open Fake.Core
+open Fake.IO
 open System.IO
 
 // Compat