+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 #### 5.0.0-beta022
 * FAKE5: New module "ChangeWatcher"
+* FAKE5: Reduce number of modules by combining some modules:
+  - `Fake.Core.BuildServer` moved to `Fake.Core.Environment`
+  - `Fake.Core.Globbing` moved to `Fake.IO.FileSystem` (includes the namespace change from `Fake.Core` to `Fake.IO`)
 
 #### 5.0.0-beta021
 * BUGFIX: Backslashes in paths wheren't escaped in intellisense.fsx - https://github.com/fsprojects/Paket/issues/3093